requirements.txt
Pillow==8.1.0
nsfw==0.3.2

